Reduce duplicate codes using macros
