+++ utils/src/types/hint.rs
@@ -1,45 +1,8 @@
 use crate::buffer::Buffer;
 use crate::decode::*;
+use crate::fixed_bytes;
 
-#[cfg_attr(test, derive(Debug, PartialEq))]
-#[derive(Default)]
-pub struct Hint([u8; 4]);
-
-impl Hint {
-    const ENCODED_LENGTH: usize = 4;
-
-    pub fn from_bytes(bytes: [u8; 4]) -> Self {
-        Hint(bytes)
-    }
-}
-
-impl RawDecoder for Hint {
-    fn step_size(&self) -> u16 {
-        1
-    }
-
-    fn decode<'a>(
-        &mut self,
-        buffer: &mut Buffer<'a>,
-        stage: &DecodeStage,
-    ) -> DecodeResult<DecodeStage> {
-        let remain = Hint::ENCODED_LENGTH - (stage.index as usize);
-        let mut idx: usize = 0;
-        while !buffer.is_empty() && idx < remain {
-            self.0[(stage.index as usize) + idx] = buffer.next_byte().unwrap();
-            idx += 1;
-        }
-        let new_index = (stage.index as usize) + idx;
-        if new_index == Hint::ENCODED_LENGTH {
-            Ok(DecodeStage::COMPLETE)
-        } else {
-            Ok(DecodeStage {
-                step: stage.step,
-                index: new_index as u16,
-            })
-        }
-    }
-}
+fixed_bytes!(Hint, 4);
 
 #[cfg(test)]
 mod tests {
