+++ utils/src/types/public_key.rs
@@ -1,51 +1,8 @@
 use crate::buffer::Buffer;
 use crate::decode::*;
+use crate::fixed_bytes;
 
-#[cfg_attr(test, derive(Debug))]
-#[derive(PartialEq)]
-pub struct PublicKey(pub [u8; 33]);
-
-impl PublicKey {
-    const ENCODED_LENGTH: usize = 33;
-
-    pub fn from_bytes(bytes: [u8; 33]) -> Self {
-        PublicKey(bytes)
-    }
-}
-
-impl Default for PublicKey {
-    fn default() -> Self {
-        PublicKey([0; 33])
-    }
-}
-
-impl RawDecoder for PublicKey {
-    fn step_size(&self) -> u16 {
-        1
-    }
-
-    fn decode<'a>(
-        &mut self,
-        buffer: &mut Buffer<'a>,
-        stage: &DecodeStage,
-    ) -> DecodeResult<DecodeStage> {
-        let remain = PublicKey::ENCODED_LENGTH - (stage.index as usize);
-        let mut idx: usize = 0;
-        while !buffer.is_empty() && idx < remain {
-            self.0[(stage.index as usize) + idx] = buffer.next_byte().unwrap();
-            idx += 1;
-        }
-        let new_index = (stage.index as usize) + idx;
-        if new_index == PublicKey::ENCODED_LENGTH {
-            Ok(DecodeStage::COMPLETE)
-        } else {
-            Ok(DecodeStage {
-                step: stage.step,
-                index: new_index as u16,
-            })
-        }
-    }
-}
+fixed_bytes!(PublicKey, 33);
 
 #[cfg(test)]
 mod tests {
